requirements.txt
fastapi[standard]
cryptography
python-dotenv